Restructure of all vue files so they are testable
