+++ src/main/resources/vue/views/login-page.vue
@@ -1,5 +1,5 @@
 <template>
-    <script type="module" src="components/login-buttons.vue"></script>
+    <script type="module" src="components/login-page-component.vue"></script>
 </template>
 <script>
     Vue.component("login-page",{template : "#login-buttons"});
